Predict likelihood of a liability customer buying personal loans
Background :
This case is about a bank which has growing customer base. The bank is interested in expanding the customer base rapidly and wants to convert majority of the liability customer to asset customer to earn Interest. Last year campaign showed a significant conversion rate over 9% .

Objective:
The department wants to build a model that will help them identify the potential customers who have higher probability of purchasing the loan.

Data :
The data shared by the bank has the information of the 5000 customers including demographic information, customer's relationship with the bank and customer response to the last personal loan campaign.